+++ src/main/java/tn/esprit/eventsproject/repositories/EventRepository.java
@@ -7,12 +7,10 @@ import tn.esprit.eventsproject.entities.Tache;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> { // Changer Integer par Long
 
     Event findByDescription(String description);
     List<Event> findByDateDebutBetween(LocalDate date1, LocalDate date2);
 
     List<Event> findByParticipants_NomAndParticipants_PrenomAndParticipants_Tache(String nom, String prenom, Tache tache);
-
-
 }
