+++ src/main/java/tn/esprit/eventsproject/repositories/ParticipantRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.eventsproject.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.eventsproject.entities.Participant;
 
-public interface ParticipantRepository extends JpaRepository<Participant, Integer> {
+public interface ParticipantRepository extends JpaRepository<Participant, Long> { // Changer Integer par Long
 }
